weston: Enable the deprecated wl_shell protocol
This is needed for the displbe which still uses wl_shell to work
with Weston 10. Otherwise there is no guest output on the screen although
both Weston and displbe don't complain about anything and both think that
everything is ok at the runtime (no errors were observed).

Just a note appears during boot which could be easily go unnoticed:
"[14:37:23.468] Note: support for the deprecated wl_shell interface is disabled.
If a legacy client still needs it, it can be re-enabled by passing
-Ddeprecated-wl-shell=true to Meson when building Weston."

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>