u-boot-android.bb: Fix the build of the u-boot-android
After switching to the Renesas BSP 5.9.0 the u-boot license checksum,
which is specified in the 'meta-renesas' layer stopped to fit the
license file in the xen-troops->u-boot repository on the Github.

This patch adjusts the used u-boot checksum.

Besides that, the patch fixes the name of the doma UBOOT_CONFIG
variable, which was accidentally changed in the wrong way, during the
adaptation to the new Yocto syntax.

Both issues were causing a build failure of the Dom0, when built with
ENABLE_ANDROID parameter.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>