u-boot-android.bb: Switch to the Android U-Boot
For the virtio product we've decided to switch to the 'master' branch of
the Android U-boot:
https://android.googlesource.com/platform/external/u-boot/

The motivation for that is obvious. It is much more suitable to use
the Android U-Boot for the 'u-boot-android' target instead of the
vanilla one.

It will allow us to use the most recent Android U-Boot features, e.g.
vendor boot image. Also, it will require less efforts to fetch the
latest Android changes in the future and will simplify our
contribution.

The recent Android U-Boot is built with Bazel instead of the build.sh.
That has caused us to rewrite the u-boot-android recipe to allow the
building approach, selected by the Android.

This patch is introducing such a change in the corresponding recipe.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>