Automate start of QEMU with DomA/DomU
This patch is intended to automate the following sequence:

- Dom0: Start DomA/DomU, pause it and notify DomD
- DomD: Fetch notification, start QEMU and notify Dom0
- Dom0: Unpause DomA/DomU

For that purpose the following things were done:

- 'doma.service' was replaced with 'doma-create.service' and
'doma-unpause.service' in Dom0
- 'domu.service' was replaced with 'domu-create.service' and
'domu-unpause.service' in Dom0
- 'backend-ready.@service' systemd template was refactored in order to
notify not only about availability of the resource, but also about its
absence.
- 'qemu-command'handler' service was added to DomD
- Other minor things was changed in order to integrate new services

In general, the desired behavior should be covered by extension of
Xen. But, as a short-term solution, it was decided to go ahead with
the systemd-level implementation.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>