Automate start of QEMU with DomA/DomU
Another thing which we need to do in order to pass the CTS is to be able
to survive the reboot of the device, as such operation takes place a lot
during the CTS execution.

From the Xen point of view, it means, that we need to survive when
domain is being reboot by itself, without triggering "xl destory
DomA/DomU" or something like that.

In order to achieve that, and to preserve restart of QEMU together with
the target guest domain, the following things were done:

- 'domu-restart-monitor' service was introduced
- Added dependency of the 'domu-create' service from the
'domu-restart-monitor' service
- 'doma-restart-monitor' service was introduced
- Added dependency of the 'doma-create' service from the
'doma-restart-monitor' service
- Extended the 'doma-create-ExecStop.sh' script
- Extended the 'domu-create-ExecStop.sh' script
- Added 'doma-create-ExecStop.sh' as 'OnFailure' action for the
'doma-restart-monitor' service
- Added 'domu-create-ExecStop.sh' as 'OnFailure' action for the
'domu-restart-monitor' service

On top of that the following minor fixes took place:

- Added usage of the vhost-vsock-pci device to QEMU
- Added dependency to 'domu-create.service' service on the
'domd' service to preserve start order of the domains. DomD should start
before DomU.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>